@@ -31,7 +31,11 @@ import { SupplierComponent } from './supplier/supplier.component';
   standalone: true,
   imports: [
     CommonModule,
-    HttpClientModule,
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule,
     TableModule,
     ButtonModule,
     DialogModule,
